a small bug fix was carried out and new tests were written
the problem with caching lists has been fixed; the problem with the accuracy of the double tail when calculating total consumption has been fixed; repository tests and  tests using selenium have been written; other minor inaccuracies have been fixed.

+++ src/main/java/ru/daniil/NauJava/entity/Meal.java
@@ -86,9 +86,9 @@ public class Meal {
         this.eatenAt = eatenAt;
     }
 
-    public Integer getTotalCalories() {
+    public Double getTotalCalories() {
         return mealEntries.stream()
-                .mapToInt(MealEntry::getCalculatedCalories)
+                .mapToDouble(MealEntry::getCalculatedCalories)
                 .sum();
     }
